♻️ [Refactor]: useStudyRoomQueries.ts 관심사에 맞게 분리
